requirements.txt
aiogram==3.4.1
openai>=1.30.0
aiosqlite==0.19.0
python-dotenv==1.0.0
tenacity==8.2.3
